requirements.txt
streamlit
mtcnn
deepface
opencv-python
pillow
pandas
tensorflow
retina-face
